+++ src/browser/js/app/widgets/plots/led.js
@@ -1,5 +1,6 @@
 var {mapToScale} = require('../utils'),
-    _widgets_base = require('../common/_widgets_base')
+    _widgets_base = require('../common/_widgets_base'),
+    {widgetManager} = require('../../managers')
 
 module.exports = class Led extends _widgets_base {
 
@@ -19,6 +20,10 @@ module.exports = class Led extends _widgets_base {
             color:'auto',
             css:'',
 
+            _plot:'plot',
+
+            widgetId:'',
+            
             _osc:'osc',
 
             range:{min:0,max:1},
@@ -43,10 +48,31 @@ module.exports = class Led extends _widgets_base {
 
         this.led = this.widget.find('span')
 
+        if (widgetData.widgetId.length) $('body').on(`sync.${this.hash}`,this.syncHandle.bind(this))
+
         this.setValue(widgetData.range.min)
 
     }
 
+    onRemove() {
+
+        $('body').off(`sync.${this.hash}`)
+
+    }
+
+    syncHandle(e) {
+
+        if (this.widgetData.widgetId!=e.id || !widgetManager.getWidgetById(e.id).length) return
+        var widget = widgetManager.getWidgetById(e.id),
+            value
+        for (var i=widget.length-1; i>=0; i--) {
+            if (widget[i].getValue) {
+                this.setValue(widget[i].getValue())
+                return
+            }
+        }
+
+    }
     setValue(v) {
 
         if (typeof v != 'number') return