+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - widgets
   - `panel/modal`: add `layout` option ([Visual Format Language](https://github.com/IjzerenHein/autolayout.js))
   - `css` option now interprets `&` as a selector for the widget's element itself (useful for targetting pseudo-elements such as `&:after`)
+  - `plots`: add `widgetId` option to `text`, `meter` and `led`
 
 ## 0.17.1